requirements.txt
aioredis==1.2.0
asgiref==3.1.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.3
django-channels==0.7.0
django-cors-headers==3.0.2
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.2
docker==4.0.2
gunicorn==19.9.0
hiredis==1.0.0
html5lib==1.0b10
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
lassie==0.11.7
lxml==4.3.4
msgpack==0.6.1
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
oauthlib==3.0.2
pdf2image==1.6.0
pdfminer.six==20181108
Pillow==6.1.0
protobuf==3.8.0
psycopg2==2.8.3
pycparser==2.19
pycryptodome==3.8.2
PyHamcrest==1.9.0
PyMySQL==0.9.3
python-oembed==0.2.4
pytz==2019.1
requests==2.18.4
requests-oauthlib==1.2.0
six==1.12.0
sortedcontainers==2.1.0
sqlparse==0.3.0
txaio==18.8.1
urllib3==1.22
webencodings==0.5.1
websocket-client==0.56.0
whitenoise==4.1.3
zope.interface==4.6.0
